Move battery and device info characteristics
