+++ pi-firmware/revvy/bluetooth/ble_characteristics.py
@@ -21,7 +21,6 @@ from revvy.bluetooth.longmessage import (
 )
 from revvy.robot.robot import BatteryStatus
 
-from revvy.utils.device_name import get_device_name, set_device_name
 from revvy.utils.logger import get_logger
 from revvy.utils.serialize import Serialize
 
@@ -192,67 +191,6 @@ class ReadOnlyCharacteristic(Characteristic):
         super().__init__({"uuid": uuid, "properties": ["read"], "value": value})
 
 
-# These are standard BLE characteristics, so we don't set a custom descriptor string for them
-class SerialNumberCharacteristic(ReadOnlyCharacteristic):
-    def __init__(self, serial: str) -> None:
-        super().__init__("2A25", serial.encode())
-
-
-class ManufacturerNameCharacteristic(ReadOnlyCharacteristic):
-    def __init__(self, name: bytes) -> None:
-        super().__init__("2A29", name)
-
-
-class ModelNumberCharacteristic(ReadOnlyCharacteristic):
-    def __init__(self, model_no: bytes) -> None:
-        super().__init__("2A24", model_no)
-
-
-class VersionCharacteristic(Characteristic):
-    version_max_length = 20
-
-    # FIXME
-    def __init__(self, uuid, version_info) -> None:
-        super().__init__({"uuid": uuid, "properties": ["read"], "value": version_info.encode()})
-        self._version = []
-
-    def onReadRequest(self, offset, callback) -> None:
-        if offset:
-            callback(Characteristic.RESULT_ATTR_NOT_LONG, None)
-        else:
-            callback(Characteristic.RESULT_SUCCESS, self._version)
-
-    def updateValue(self, version) -> None:
-        if len(version) > self.version_max_length:
-            version = version[: self.version_max_length]
-        self._version = version.encode("utf-8")
-
-
-class SystemIdCharacteristic(Characteristic):
-    def __init__(self) -> None:
-        super().__init__({"uuid": "2A23", "properties": ["read", "write"]})
-
-    def onReadRequest(self, offset, callback) -> None:
-        if offset:
-            callback(Characteristic.RESULT_ATTR_NOT_LONG, None)
-        else:
-            callback(Characteristic.RESULT_SUCCESS, get_device_name().encode("utf-8"))
-
-    def onWriteRequest(self, data: bytes, offset, withoutResponse, callback) -> None:
-        if offset:
-            callback(Characteristic.RESULT_ATTR_NOT_LONG)
-        else:
-            try:
-                name = data.decode("ascii")
-                if 0 < len(name) <= 15:
-                    set_device_name(name)
-                    callback(Characteristic.RESULT_SUCCESS)
-                else:
-                    callback(Characteristic.RESULT_UNLIKELY_ERROR)
-            except UnicodeDecodeError:
-                callback(Characteristic.RESULT_UNLIKELY_ERROR)
-
-
 class MobileToBrainFunctionCharacteristic(Characteristic):
     def __init__(self, uuid, min_length, max_length, description: bytes, callback) -> None:
         self._callbackFn = callback
@@ -339,40 +277,3 @@ class LongMessageCharacteristic(Characteristic):
             log(traceback.format_exc())
         finally:
             callback(result)
-
-
-class UnifiedBatteryInfoCharacteristic(Characteristic):
-    def __init__(self, uuid, description, initial: BatteryStatus) -> None:
-        super().__init__(
-            {
-                "uuid": uuid,
-                "properties": ["read", "notify"],
-                "descriptors": [Descriptor({"uuid": "2901", "value": description})],
-            }
-        )
-
-        self._value = initial
-
-    def onReadRequest(self, offset, callback) -> None:
-        if offset:
-            callback(Characteristic.RESULT_ATTR_NOT_LONG, None)
-        else:
-            callback(
-                Characteristic.RESULT_SUCCESS,
-                [
-                    self._value.main,
-                    self._value.chargerStatus,
-                    self._value.motor,
-                    self._value.motor_battery_present,
-                ],
-            )
-
-    def updateValue(self, new_value: BatteryStatus) -> None:
-        if new_value == self._value:
-            return
-
-        self._value = new_value
-
-        update_notified_value = self.updateValueCallback
-        if update_notified_value:
-            update_notified_value(self._value)
